refactor: 아바타 선택 페이지와 관련된 Dto를 하나의 클래스로 관리
