+++ src/main/java/com/fastcampus/finalproject/controller/ProjectController.java
@@ -1,13 +1,12 @@
 package com.fastcampus.finalproject.controller;
 
 import com.fastcampus.finalproject.config.security.AuthUtil;
+import com.fastcampus.finalproject.dto.AvatarDto.*;
 import com.fastcampus.finalproject.dto.ResponseWrapper;
 import com.fastcampus.finalproject.dto.request.InsertTextPageRequest;
 import com.fastcampus.finalproject.dto.response.CreateProjectResponse;
-import com.fastcampus.finalproject.dto.response.GetAvatarPageResponse;
 import com.fastcampus.finalproject.dto.response.GetHistoryResponse;
 import com.fastcampus.finalproject.dto.response.GetTextPageResponse;
-import com.fastcampus.finalproject.dto.request.GetAvatarPreviewRequest;
 import com.fastcampus.finalproject.dto.response.*;
 import com.fastcampus.finalproject.service.ProjectService;
 import lombok.RequiredArgsConstructor;
@@ -38,13 +37,13 @@ public class ProjectController {
     }
 
     @GetMapping("/projects/{projectId}/avatar")
-    public ResponseWrapper<GetAvatarPageResponse> getAvatarSelectionPage(@PathVariable Long projectId) {
+    public ResponseWrapper<AvatarPageResponse> getAvatarSelectionPage(@PathVariable Long projectId) {
         return new ResponseWrapper<>(projectService.getAvatarPageData(AuthUtil.getCurrentUserUid(), projectId))
                 .ok();
     }
 
     @PostMapping("/projects/avatar-preview")
-    public ResponseWrapper<GetAvatarPreviewResponse> getAvatarPreview(@RequestBody GetAvatarPreviewRequest request) {
+    public ResponseWrapper<AvatarPreviewResponse> getAvatarPreview(@RequestBody AvatarPageRequest request) {
         return new ResponseWrapper<>(projectService.getAvatarPreview(request))
                 .ok();
     }
