--- src/main/java/com/fastcampus/finalproject/dto/request/GetAvatarPreviewRequest.java
@@ -1,11 +1,0 @@
-package com.fastcampus.finalproject.dto.request;
-
-import lombok.Data;
-
-@Data
-public class GetAvatarPreviewRequest {
-
-    private String avatarName;
-    private String avatarType;
-    private String bgName;
-}
